more lobby fixing, eventually will deviate player 0 from others
